# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Availability": {
        "properties": {
          "since": {
            "format": "date-time",
            "title": "Since",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AvailabilityStatus"
          }
        },
        "required": [
          "status",
          "since"
        ],
        "title": "Availability",
        "type": "object"
      },
      "AvailabilityStatus": {
        "enum": [
          "available",
          "paused"
        ],
        "title": "AvailabilityStatus",
        "type": "string"
      },
...
      "CaseStatus": {
        "description": "Stored state machine of a case (see ``Case``).",
        "enum": [
          "routing",
          "queued",
          "assigned",
          "in_progress",
          "in_call",
          "to_call",
          "awaiting_approval",
          "closed"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
...
      "ProblemCode": {
        "enum": [
          "invalid_credentials",
          "mfa_invalid",
          "mfa_challenge_invalid",
          "account_locked",
          "unauthenticated",
          "session_expired",
          "forbidden",
          "not_found",
          "method_not_allowed",
          "conflict",
          "concurrent_update",
          "invalid_transition",
          "case_not_assigned",
          "case_closed",
          "channel_not_supported",
          "idempotency_conflict",
          "invalid_value",
          "policy_violation",
          "validation_error",
          "domain_error",
          "application_error",
          "invalid_topic",
          "invalid_message",
          "http_error",
          "internal_error"
        ],
        "title": "ProblemCode",
        "type": "string"
      },
      "ProblemDetails": {
        "additionalProperties": true,
        "description": "RFC 7807 problem. ``code`` is the stable machine identifier clients branch on.\n\nThe optional members below are the documented extensions; a domain error may add other\nstructured details (e.g. ``channel`` on ``channel_not_supported``), hence\n``additionalProperties``.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "currentStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseStatus"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "invalid_transition, case_closed: the case status now."
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "validation_error: one entry per invalid field.",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "remainingAttempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "invalid_credentials, mfa_invalid: failed attempts left before the lock.",
            "title": "Remainingattempts"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Requestid"
          },
          "requiredRoles": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StaffRole"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "forbidden: roles that may perform the action.",
            "title": "Requiredroles"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "examples": [
              "urn:cc-platform:problem:account_locked"
            ],
            "title": "Type",
            "type": "string"
          },
          "unlockAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "account_locked: when the lock ends.",
            "title": "Unlockat"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "StaffRole": {
        "description": "Platform roles. They combine: one person may hold several (brief §1).",
        "enum": [
          "analyst",
          "supervisor",
          "automation",
          "admin"
        ],
        "title": "StaffRole",
        "type": "string"
      },
...
      "UpdateAvailabilityRequest": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/AvailabilityStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "UpdateAvailabilityRequest",
        "type": "object"
      },
      "ValidationIssue": {
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      }
    },
    "securitySchemes": {
      "CustomerToken": {
        "description": "Customer token returned by POST /api/v1/customer/sessions (simulator).",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionToken": {
        "description": "Session token returned by POST /api/v1/auth/mfa.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API of the LATAM Bank contact-center platform (transaction-dispute intake). Errors are RFC 7807 `application/problem+json` with a stable `code`. Realtime: WebSocket `/api/v1/ws?token=…` with `{type,id,occurredAt,data}` envelopes.",
    "title": "CC Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/availability": {
      "get": {
        "operationId": "availability_get_availability",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "My availability (never set = paused)",
        "tags": [
          "availability"
        ]
      },
      "put": {
        "operationId": "availability_set_availability",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvailabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Set my availability (same status = no change, no event)",
        "tags": [
          "availability"
        ]
      }
    },
...
  }
}